Add NetStandard to CanOSPerformPKeyAuth. Refactor duplicate code into DeviceAuthHelper.
