+++ src/client/Microsoft.Identity.Client/Platforms/netdesktop/NetDesktopDeviceAuthManager.cs
@@ -3,22 +3,13 @@
 
 using System;
 using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
-using System.Net;
 using System.Net.Http.Headers;
-using System.Security;
 using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
-using System.Security.Permissions;
-using System.Threading.Tasks;
-using Microsoft.Identity.Client.Http;
 using Microsoft.Identity.Client.OAuth2;
 using Microsoft.Identity.Client.PlatformsCommon.Interfaces;
 using Microsoft.Identity.Client.PlatformsCommon.Shared;
 using Microsoft.Identity.Client.Utils;
-using Microsoft.Win32.SafeHandles;
-using Microsoft.Identity.Client.Platforms.net45.Native;
 
 namespace Microsoft.Identity.Client.Platforms.net45
 {
@@ -27,13 +18,13 @@ namespace Microsoft.Identity.Client.Platforms.net45
         public bool TryCreateDeviceAuthChallengeResponseAsync(HttpResponseHeaders responseHeaders, Uri endpointUri, out string responseHeader)
         {
             responseHeader = string.Empty;
-            string authHeaderTemplate = "PKeyAuth {0}, Context=\"{1}\", Version=\"{2}\"";
             X509Certificate2 certificate = null;
 
             if (!DeviceAuthHelper.IsDeviceAuthChallenge(responseHeaders))
             {
                 return false;
             }
+
             if (!DeviceAuthHelper.CanOSPerformPKeyAuth())
             {
                 responseHeader = DeviceAuthHelper.GetBypassChallengeResponse(responseHeaders);
@@ -49,7 +40,7 @@ namespace Microsoft.Identity.Client.Platforms.net45
 
             try
             {
-                certificate = FindCertificate(challengeData);
+                certificate = DeviceAuthHelper.FindCertificate(challengeData);
             }
             catch (MsalException ex)
             {
@@ -71,79 +62,9 @@ namespace Microsoft.Identity.Client.Platforms.net45
                 sig = rsa.SignData(responseJWT.GetResponseToSign().ToByteArray());
             }
 
-            string signedJwt = string.Format(CultureInfo.InvariantCulture, "{0}.{1}", responseJWT.GetResponseToSign(),
-                Base64UrlHelpers.Encode(sig));
-            string authToken = string.Format(CultureInfo.InvariantCulture, " AuthToken=\"{0}\"", signedJwt);
-
-            responseHeader = string.Format(CultureInfo.InvariantCulture, authHeaderTemplate, authToken,
-                challengeData["Context"],
-                challengeData["Version"]);
+            DeviceAuthHelper.FormatResponseHeader(responseJWT, sig, challengeData, out responseHeader);
 
             return true;
-        }
-
-        private static X509Certificate2 FindCertificate(IDictionary<string, string> challengeData)
-        {
-            var store = new X509Store(StoreName.My, StoreLocation.CurrentUser);
-            try
-            {
-                store.Open(OpenFlags.ReadOnly);
-                var certCollection = store.Certificates;
-                if (challengeData.ContainsKey("CertAuthorities"))
-                {
-                    return FindCertificateByCertAuthorities(challengeData, certCollection);
-                }
-
-                X509Certificate2Collection signingCert = null;
-                signingCert = certCollection.Find(X509FindType.FindByThumbprint, challengeData["CertThumbprint"],
-                    false);
-                if (signingCert.Count == 0)
-                {
-                    throw new MsalException(MsalError.DeviceCertificateNotFound,
-                        string.Format(CultureInfo.CurrentCulture, MsalErrorMessage.DeviceCertificateNotFoundTemplate,
-                            "Cert thumbprint:" + challengeData["CertThumbprint"]));
-                }
-
-                return signingCert[0];
-            }
-            finally
-            {
-                store.Close();
-            }
-        }
-
-        private static X509Certificate2 FindCertificateByCertAuthorities(IDictionary<string, string> challengeData, X509Certificate2Collection certCollection)
-        {
-            X509Certificate2Collection signingCert = null;
-            string[] certAuthorities = challengeData["CertAuthorities"].Split(new[] { ";" },
-                StringSplitOptions.None);
-            foreach (var certAuthority in certAuthorities)
-            {
-                //reverse the tokenized string and replace "," with " + "
-                string[] dNames = certAuthority.Split(new[] { "," }, StringSplitOptions.None);
-                string distinguishedIssuerName = dNames[dNames.Length - 1];
-                for (int i = dNames.Length - 2; i >= 0; i--)
-                {
-                    distinguishedIssuerName += " + " + dNames[i].Trim();
-                }
-
-                signingCert = certCollection.Find(X509FindType.FindByIssuerDistinguishedName,
-                    distinguishedIssuerName, false);
-                if (signingCert.Count > 0)
-                {
-                    break;
-                }
-            }
-
-            if (signingCert == null || signingCert.Count == 0)
-            {
-                throw new MsalException(MsalError.DeviceCertificateNotFound,
-                    string.Format(CultureInfo.CurrentCulture,
-                        MsalErrorMessage.DeviceCertificateNotFoundTemplate,
-                        "Cert Authorities:" + challengeData["CertAuthorities"]));
-            }
-
-            return signingCert[0];
         }
     }
 }
