+++ src/client/Microsoft.Identity.Client/PlatformsCommon/Shared/DeviceAuthHelper.cs
@@ -5,12 +5,10 @@ using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
-using System.Net;
 using System.Net.Http.Headers;
-using System.Runtime.InteropServices;
-using Microsoft.Identity.Client.Http;
+using System.Security.Cryptography.X509Certificates;
+using Microsoft.Identity.Client.OAuth2;
 using Microsoft.Identity.Client.Utils;
-using Microsoft.Win32;
 
 namespace Microsoft.Identity.Client.PlatformsCommon.Shared
 {
@@ -77,11 +75,99 @@ namespace Microsoft.Identity.Client.PlatformsCommon.Shared
         //This corresponds to windows 7, 8, 8.1 and their server equivalents.       
         public static bool CanOSPerformPKeyAuth()
         {
-#if NET_CORE || NET5_WIN || DESKTOP
+#if NET_CORE || NET5_WIN || DESKTOP || NETSTANDARD
             return !Platforms.Features.DesktopOs.DesktopOsHelper.IsWin10OrServerEquivalent();
 #else
             return false;
 #endif
+        }
+
+        /// <summary>
+        /// Formats PKeyAuth header
+        /// </summary>
+        public static void FormatResponseHeader(
+            DeviceAuthJWTResponse responseJWT,
+            byte[] signedResponse,
+            IDictionary<string, string> challengeData,
+            out string responseHeader)
+        {
+            string authHeaderTemplate = "PKeyAuth {0}, Context=\"{1}\", Version=\"{2}\"";
+
+            string signedJwt = string.Format(CultureInfo.InvariantCulture, "{0}.{1}", responseJWT.GetResponseToSign(), Base64UrlHelpers.Encode(signedResponse));
+            string authToken = string.Format(CultureInfo.InvariantCulture, " AuthToken=\"{0}\"", signedJwt);
+
+            responseHeader = string.Format(CultureInfo.InvariantCulture, authHeaderTemplate, authToken,
+                challengeData["Context"],
+                challengeData["Version"]);
+        }
+
+        public static X509Certificate2 FindCertificate(IDictionary<string, string> challengeData)
+        {
+            var store = new X509Store(StoreName.My, StoreLocation.CurrentUser);
+            try
+            {
+                store.Open(OpenFlags.ReadOnly);
+                var certCollection = store.Certificates;
+                if (challengeData.ContainsKey("CertAuthorities"))
+                {
+                    return FindCertificateByCertAuthorities(challengeData, certCollection);
+                }
+
+                X509Certificate2Collection signingCert = null;
+                signingCert = certCollection.Find(X509FindType.FindByThumbprint, challengeData["CertThumbprint"],
+                    false);
+                if (signingCert.Count == 0)
+                {
+                    throw new MsalException(MsalError.DeviceCertificateNotFound,
+                        string.Format(CultureInfo.CurrentCulture, MsalErrorMessage.DeviceCertificateNotFoundTemplate,
+                            "Cert thumbprint:" + challengeData["CertThumbprint"]));
+                }
+
+                return signingCert[0];
+            }
+            finally
+            {
+
+#if NETSTANDARD
+                store.Dispose();
+#else
+                store.Close();
+#endif
+            }
+        }
+
+        private static X509Certificate2 FindCertificateByCertAuthorities(IDictionary<string, string> challengeData, X509Certificate2Collection certCollection)
+        {
+            X509Certificate2Collection signingCert = null;
+            string[] certAuthorities = challengeData["CertAuthorities"].Split(new[] { ";" },
+                StringSplitOptions.None);
+            foreach (var certAuthority in certAuthorities)
+            {
+                //reverse the tokenized string and replace "," with " + "
+                string[] dNames = certAuthority.Split(new[] { "," }, StringSplitOptions.None);
+                string distinguishedIssuerName = dNames[dNames.Length - 1];
+                for (int i = dNames.Length - 2; i >= 0; i--)
+                {
+                    distinguishedIssuerName += " + " + dNames[i].Trim();
+                }
+
+                signingCert = certCollection.Find(X509FindType.FindByIssuerDistinguishedName,
+                    distinguishedIssuerName, false);
+                if (signingCert.Count > 0)
+                {
+                    break;
+                }
+            }
+
+            if (signingCert == null || signingCert.Count == 0)
+            {
+                throw new MsalException(MsalError.DeviceCertificateNotFound,
+                    string.Format(CultureInfo.CurrentCulture,
+                        MsalErrorMessage.DeviceCertificateNotFoundTemplate,
+                        "Cert Authorities:" + challengeData["CertAuthorities"]));
+            }
+
+            return signingCert[0];
         }
     }
 }
